Updates for ubuntu:17.10 and SDK tools
The repos for `ubuntu:17.04` have been moved and this causes issues with all the package managers (see [this](https://ubuntuforums.org/showthread.php?t=2382832)). Updating the image to `17.10` seems to be the better idea.
Google has released a new version of SDK tools and I don't see why one wouldn't want to update to this.